# apps/backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "完整的 POS 系統後端 API，包含商品管理、分類管理、庫存管理等功能",
        "title": "POS System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@pos.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/api",
    "paths": {
...
        "/product-categories": {
            "get": {
                "description": "取得所有商品分類，支援 active 和 parent_id 篩選",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "取得商品分類列表",
                "parameters": [
                    {
                        "type": "string",
                        "default": "true",
                        "description": "是否只顯示啟用的分類 (true/false)",
                        "name": "active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "父分類 ID；不傳回全部分類，空字串表示頂層分類，UUID 表示指定父分類下的子分類",
                        "name": "parent_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/dto.CategoryListResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "新增一個新的商品分類。頂層分類請省略 parent_id 或傳 null；子分類請傳父分類 UUID；parent_id 不接受空字串。order 不傳時會自動排到同層最後。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "新增商品分類",
                "parameters": [
                    {
                        "description": "分類資訊",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCategorySuccessResponse"
                        }
                    },
                    "400": {
                        "description": "參數錯誤或分類名稱已存在",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/product-categories/{id}": {
            "delete": {
                "description": "刪除指定分類（會一併刪除子分類，商品轉為未分類）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "刪除商品分類",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分類 ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteCategorySuccessResponse"
                        }
                    },
                    "400": {
                        "description": "分類仍有庫存商品",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "分類不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CategoryListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CategoryResponse"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "total": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "2c619823-aa6b-4e7e-931f-c8358bb07861"
                },
                "name": {
                    "type": "string",
                    "example": "上衣"
                },
                "order": {
                    "type": "integer",
                    "example": 1
                },
                "parent_id": {
                    "type": "string",
                    "x-nullable": true,
                    "example": "2c619823-aa6b-4e7e-931f-c8358bb07861"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "dto.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "上衣"
                },
                "order": {
                    "type": "integer",
                    "example": 1
                },
                "parent_id": {
                    "type": "string",
                    "x-nullable": true,
                    "example": "2c619823-aa6b-4e7e-931f-c8358bb07861"
                }
            }
        },
        "dto.CreateCategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                },
                "message": {
                    "type": "string",
                    "example": "分類新增成功"
                }
            }
        },
        "dto.CreateCategorySuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.CreateCategoryResponse"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.DeleteCategoryResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "分類刪除成功"
                }
            }
        },
        "dto.DeleteCategorySuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.DeleteCategoryResponse"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "failed_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "6d1fb22a-2eb3-4a8a-9750-b4d63b56b971"
                    ]
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ValidationFieldError"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "請求參數驗證失敗"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorDetail"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.ValidationFieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "image_ids[0]"
                },
                "message": {
                    "type": "string",
                    "example": "必須是合法 UUID"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}